修改积分类型,测试部分QPS,更新客户端配置

+++ dao/redis/user.go
@@ -139,7 +139,7 @@ func (c *UserRDB) ExistsSameID(ctx context.Context, userID string) bool {
 }
 
 // 更新积分
-func (c *UserRDB) UpdatePoints(ctx context.Context, userID string, num int, points_type points.PointsType) error {
+func (c *UserRDB) UpdatePoints(ctx context.Context, userID string, num float32, points_type points.PointsType) error {
 	user, err := c.GetByID(ctx, userID)
 	if err != nil {
 		return err
